6.2 - Fixed some issues in bank. Passing checkout to concentrator. Possible changes and refactoring.

+++ Bank/src/main/java/com/ftn/model/dto/onlinepayment/PaymentInquiryDTO.java
@@ -27,7 +27,7 @@ public class PaymentInquiryDTO {
     private double amount;
 
     @NotNull
-    private int merchantOrderId;
+    private Long merchantOrderId;
 
     @NotNull
     @JsonFormat(pattern="yyyy-MM-dd")
